EIICS/requirements.txt
pillow>=9.0.0
pycryptodome>=3.15.0